:ok_hand: Optimization by splitting modules & add lazy-loading
