+++ src/app/meal-list/meal/meal-info/meal-info.component.scss
@@ -1,4 +1,4 @@
-@import "../../../assets/styles/variables";
+@import "../../../../assets/styles/variables";
 
 .meal-info-wrapper {
   height: 100%;
