+++ src/app/meal-list/meal/meal-info/meal-info.component.ts
@@ -1,5 +1,5 @@
 import {Component, Input, OnInit} from '@angular/core';
-import {RecipeModel} from "../../models/recipe.model";
+import {RecipeModel} from "../../../models/recipe.model";
 
 @Component({
   selector: 'app-meal-info',
