+++ src/app/meal-list/meal/meal-picture/meal-picture.component.scss
