+++ src/app/meal-list/meal/meal.component.scss
@@ -1,4 +1,4 @@
-@import "../../assets/styles/variables";
+@import "../../../assets/styles/variables";
 
 .meal {
   display: flex;
